Bugfix and dump urllib in favor of requests
Don't reference Bittle model, but use self.model
---

This is very likely the cause of the MultipleObjectsExist bug, as
referencing a manager on a model from that manager allows Python to
parallize.

Dump urllib
---

It is braindead and always has been. Once again shown by the fact that
it wants POST data in bytes objects.